Update genatic_algorithm_method.py (optimal_energy equal minimum energy in fasta sequence)
changes in the menu of the output for better view, also if the optimal of the sequence is unkown so the optimal of the sequence will be equal the minimum energy. Moreover, an interactive mode for the user input, displays best energy and compares it with known optimal from optimal_2d.